backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv

# Environment Variables Required:
# GOOGLE_API_KEY=your_gemini_api_key_here
# ZOHO_WORKSPACE_ID=your_workspace_id_here (optional, for real data queries)
# ACCOUNTS_SERVER_URL=https://accounts.zoho.in
# ANALYTICS_SERVER_URL=https://analyticsapi.zoho.in
# ZOHO_CLIENT_ID=your_client_id_here
# ZOHO_CLIENT_SECRET=your_client_secret_here
# ZOHO_REFRESH_TOKEN=your_refresh_token_here
